Create Enemy
This script covers all the actions and the properties of the enemy. In order to get the enemy to move, we use the 'nodes' that are part of the Unity library. This smoothens the movement of the enemy. The enemies can take damage when they are hit by the bullets, and destroyed when their HP is depleted.