Rename Flip Card Animations/Main to Flip Card Animation/Main
We start with an empty scene, and we want to create everything; animations, the card models, their sprites, via C#. This is what this script here is for. We create the sizes and the number of cards we want in the scene, give them an animation, and add an event that must be triggered in order to activate the animations.