Create Register
This script gives the player to ability to register to a website. This script creates a .txt file with all the information the player has entered so they can login. The script also makes a list of all the characters that are accepted and refused by the software, as well as the minimum length of the password and the username.