Create Domino Painter
Using multiple system.serialize we can create 3D dominoes; give them a texture; their transform variables; give them the collider, rigidbody they need; by clicking on the right mouse button when the raycast touches a specific area in the scene. We also have the ability to save the scene as a new level design scene for our game, and to undo our work by pressing on the left mouse button.